[pages] setup minimum viable by-page view
This creates a pages collection, puts a thumbnail-based view of pages
on the index, and creates a per-page view with the web-scaled enhanced
image (aside from the cover), links to the full-res and enhanced
full-res image, and a next/prev link where appropriate.

This also creates collection entries for the edge-cases and one
non-edge-case pages for testing.  A subsequent commit will
auto-generate the collection entries for the remaining pages.