[chapters] create page per-chapter
This sets up a new chapters collection, with one entry per chapter.  Each such entry contains my best effort at transcribing the text on the first page of that chapter.

These chpter pages seem to be an intro for the chapters.  It's the hand-writing equivalent of the several pages of story that every recipe website seems to have before you get to the actual recipe.  Just shorter because there's no additional ad space to be gained by padding the page, and hand-writing takes much more effort than typing does.

The actual contents seems to be a mix of quotes, guidelines and opinions designed to start flame-wars.