[layout] no paragraphs around thumbnails
This is so ugly code-wise, but produces a better page layout.

It appears as though liquid for-loops on a single line do not produce <p> tags around each line whereas multi-line for loops do result in a paragraph per iterations.

This change makes the loops a lot harder to read, but also results in a noticeably better page layout.  Ugh.  Maybe I should switch from markdown to html :/