[cloudflare] upload only the non-full_res assets
The full res assets are too big to fit on the github action worker.  This sets up the single existing job to perform a sparse checkout with everything but the full_res assets.

The intent is to get this working, and then created 3 additional upload jobs, one per full-res asset type, with a sparse profile for only that asset type in full res.