[header] get rid of junk in header
This tries to set the header links to a single link to about.html but
it doesn't work.  As best I can tell, site.pages is mteant to be a list
of all pages on the site.  However, because I have a collection named
_pages site.pages gets overidden with just the members of that collection.

Subsequent diffs are going to refactor such that I can reasonably
rename that collection without breaking anything.