[layout] import favoicon stuff
NB: This doesn't actually work yet, because it depends on a
custom-head.html include not available in latest jekyll-minima release.

Subsequent changes will migrate to a version of jekyll-minima that does
suppor tthe needed include.  It's a backwards-incompatible change so
requires more than jsut changing the version in _config.yml