[images] set width attribute on all images
This is a step towards only committing the full-res images. For
cloudflare in production we're going to use the images transform
already working for the recipe pics.

In dev, we might just load
the huge full-res images and deal with the slowness or we might end
up with an image transforming plugin.  There are imagemagick bindings
for ruby, and there seems to be at least one jekyll plugin which can
resize images.