[images] use jekyll-resize for non-prod images
This stops using the committed resizes of the images.  Previous diffs
have switched production builds to using cloudflare's image resizing.

This uses the jekyll-resize plugin for dev buiilds.  This generates
(and caches for later builds) resized versions of the images.  It makes
local builds quite slow the first time, and noticeably slower for
incremental builds.

The reason we're doing it though is because it massively simplifies
the plan for adding recipe photos one at a time.  Previously we'd need
to arrange to run the image resizing scripts and commit the results
every time someone contributes a recipe photo.  With this, we just need
to commit the one photo.  We'll take care of resizing at build time in
local and runtime in prod.

Subsequent diffs are still needed to clean up the leftovers of the
committed resized images.