[R2] rework upload github action
This moves from the... inefficient r2 upload workflow we were
previously using to an rclone-based uploaf workflow, which checks to
see if a file has changed before uploading it and will also delete
files from r2 that get deleted in the repo.

Additionally it drops the non-full-res uploads, since we no longer
serve images, in either prod or dev, from those dirs.

Finally, it drops the full_res prefix in R2, as part of the process of
cleaning up the legacy, pre-generated and commited resized images.
Subsequent diffs will delete the pre-generated images, move the full-size
images to the root of assets/ and change the image link generation code
to point to the new location.