[cloudflare] hopefully fix preview deployments
The previous commit was an experiment to see what cloudflare preview
deployment did with some of the suggested env names in the wrangler
docs.  It gave me the following error:

```
10:33:57.272	✘ [ERROR] Running configuration file validation for Pages:
10:33:57.361
10:33:57.362	    - Configuration file contains the following environment names that are not supported by Pages projects:
10:33:57.362	      "staging","dev".
10:33:57.362	      The supported named-environments for Pages are "preview" and "production".
```

This strongly suggests "preview" is the env we need for the preview
deployments.  This diff drops the other ones I was experimenting with
and also the env variables whose whol;e prupose was to try to get the
env being used printed as part of the build output.

Hopefully this will work, and I can just merge the pull request.